refactor getTransactions
