+++ src/services/contractMethods.js
@@ -41,7 +41,7 @@ export function convertArgs(tokenSymbol, CUR = 'USD') {
 export async function callClient(searchValue){
   const tokenSearch = convertArgs(searchValue.toUpperCase())
   return await window.clientAcct.functionCall(
-    'client.check.testnet',
+    'client.development.testnet',
     'demo_token_price',
     {
       symbol: tokenSearch,
@@ -53,7 +53,7 @@ export async function callClient(searchValue){
 
 export async function getReceivedVal(nonce){
   return await window.clientAcct.viewFunction(
-    'client.check.testnet',
+    'client.development.testnet',
     'get_received_val',
     { nonce: nonce.toString() }
   )
