+++ src/services/utils.js
@@ -1,4 +1,5 @@
 import { connect, keyStores, KeyPair } from 'near-api-js'
+import { getBlockByID } from './contractMethods';
 
 const nearConfig = {
   networkId: 'testnet',
@@ -7,6 +8,7 @@ const nearConfig = {
   walletUrl: 'https://wallet.testnet.near.org',
   helperUrl: 'https://helper.testnet.near.org'
 };
+
 
 export async function initContract() {    
   const keyStore = new keyStores.InMemoryKeyStore()
@@ -18,3 +20,41 @@ export async function initContract() {
   window.near = near
   window.clientAcct = await near.account(`client.${process.env.ACCOUNT_ID}.testnet`)
 }
+
+export async function getTransactions(firstBlock, lastBlock){
+  const blockArr = [];
+  for (let i = firstBlock; i <= lastBlock; i++) {
+    blockArr.push(i)
+  }
+
+  const blockResults = await Promise.all(blockArr.map(block => {
+    return getBlockByID(block);
+  }))
+
+  console.log('blockResults', blockResults);
+  
+  const chunkArr = [];
+  blockResults.map(block => {
+    block.chunks.map(chunk => {
+      chunkArr.push(chunk.chunk_hash);
+    });
+  });
+  
+  const chunkDetails = await Promise.all(chunkArr.map(chunk => {
+      return window.near.connection.provider.chunk(chunk);
+  }));
+  
+  console.log('chunkDetail', chunkDetails)
+  
+  const transactions = []
+  chunkDetails.map(chunk => {
+    chunk.transactions?.map(transaction => {
+      console.log(transaction)
+      if(transaction.signer_id.includes('development.testnet')) {
+        transactions.push(transaction);
+      };
+    });
+  });
+  console.log(transactions)
+  }
+
